chore: polyfill scroll behavior
Older browsers, namely Edge 18 and earlier, don't support the use of `scrollBy` and instead throw errors.

In those cases, say when an item is being dragged from one spot to another and we are trying to automatically scroll a container, dragging will still be fine but scrolling won't happen and an error will be thrown.

Adding the polyfill here means that we now support the intended scroll behavior in those older browser cases.